make toast appear once; move skip to dropdown only
